.idea/src/requirements.txt
# 基于人脸和指纹的多模态身份识别项目 - 依赖包列表
# 核心深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# 科学计算和数据处理
numpy>=1.21.0
scikit-learn>=1.0.0

# 图像处理
pillow>=8.0.0

# 配置管理和序列化
pyyaml>=5.4.0

# 进度条和可视化
tqdm>=4.62.0
matplotlib>=3.5.0
seaborn>=0.11.0  # 用于更美观的混淆矩阵

# 实验监控和日志
tensorboard>=2.7.0

# 可选：GPU加速支持 (如果有CUDA)
# torchtext>=0.10.0  # 如果需要文本处理
# torchaudio>=0.9.0  # 如果需要音频处理